Extract DAG cycle tester
